+++ src/Http/Responses/InternalServerError.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rubix\Server\Http\Responses;
+
+class InternalServerError extends Response
+{
+    public function __construct()
+    {
+        parent::__construct(500);
+    }
+}