+++ src/Http/Responses/Response.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rubix\Server\Http\Responses;
+
+use GuzzleHttp\Psr7\Response as GuzzleResponse;
+
+class Response extends GuzzleResponse
+{
+    //
+}